+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "eq-author-api",
   "version": "1.0.0",
-  "main": "src/server.js",
+  "main": "app.js",
   "license": "MIT",
   "scripts": {
     "start": "nodemon --debug"
